+++ tox.ini
@@ -38,9 +38,11 @@ max-complexity=6
 inline-quotes=double
 ; ignore:
 ; C812 - Missing trailing comma
+; C813 - Missing trailing comma in Python 3
 ; D104 - Missing docstring in public package
 ; D400 - First line should end with a period
-ignore=C812,D104,D400
+; E203 - Whitespace before ':'
+ignore=C812,C813,D104,D400,E203
 
 [mypy]
 strict_optional=True
